pyproject.toml
[project]
name = "little-jokebook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.2",
    "google-genai>=1.63.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "temporalio>=1.22.0",
    "yt-dlp>=2026.2.4",
]

[dependency-groups]
dev = []
